[Study][DesignPattterns☀️] Factory method 패턴 init
